add hover delay to allow diagonal moves without changing category
Many newer users have less-than-ideal mouse dexterity.  Requiring a
brief pause before changing the application category is really nice.
It is even nice for power users who can mouse directly towards their
desired application without it disappearing on them when the
category changes.  With this change, using Hover as a default is
preferable.

The default of 150 requires a very brief pause. It will be too
fast for beginners, but should be acceptable for power users.  250
should be a pretty good number for newbies.