enhance search to match filename or keyword metadata
Some programs like "nautilus" don't contain the executable name
in either the application name or in the description.  Someone who
commonly starts a program from the commandline should be able
to find it from the search.

Finding a program based on the keywords can be very helpful
for new users who know what kind of program they run but don't
know the oddball name that it has.  This allows better
searching than just the description, and allows flexibility
for things like control panel to have many keywords added.