allow customizable label width for grid layout
Previously the labels in the grid were unreadable for most column
formats.  This patch allows user to enter a numeric label length, so there
is full customization by the user for any size he wants.  Reasonable
choices are offered via the settings menu, but any size can be set
using gsettings or dconf-editor.

A label size of 0 is offered to disable labels completely.  This could be
good for compact view - where the label was generally useless anyway.
The name and description of each icon still shows up in the info bar.