allow gnomenu preferences/settings button to be hidden
DEFAULT: ALL
Always show the settings button

OPTION: NONE
Never show the settings button

OPTION: <unix-group>
Only show the settings to the members of this group.  "sudo" would be
a good example for debian based systems, and would be my recommended
default except that I understand that distros like Fedora and OpenSuse
don't have a sudo group.

This is a "hidden" option since it is not included in prefs.js.
This prevents a user from mistakenly locking himself out of
gnomenu's preferences.

In order to be able to manage this via gsettings/deconf-editor, the
schema must be copied into /usr/share/glib-2.0/schemas.
Alternatively an administrator could change the default in the
extensions ./schema folder and recompile with glib-compile-schema ./schema.